refactor: fold WidgetPosition into ToolbarButtons (#1314)
Because WidgetPosition is only ever used with ToolbarButtons, and
ToolbarButtons is only ever used with WidgetPosition.

Removed the useless "mock/widget-position.js"; try not to be too
horrified by the thing I had to do to keep the
"widget-interaction-point.js" test still working: it was taking a
class component and then calling `new` on it, which is something I don't
think I've ever seen before (React takes care of instantiating
components).

Seems like "widget-interaction-point" was only partially refactored
out of existence in e8561dc78744f120c1d1b6825b697310ddd38b91 ("Initial
refactor for AlloyEditor 2.0", February 2018), but a number of
straggling references to it remained. With this commit, the only
remaining reference to it is in the test filename, which I will fix in
the next commit.
