--- test/ui/test/mock/widget-position.js
@@ -1,5 +1,0 @@
-import WidgetPosition from '../../../../src/components/base/widget-position';
-
-class Component {}
-
-export default WidgetPosition(Component);
